Implement comment marker functionality in GlobalAudioManager
- Added methods to render, update, and clear comment markers in the audio waveform.
- Integrated event listeners for comment updates and keyboard shortcuts to add comments at the current playback position.
- Enhanced the user experience by providing visual feedback for resolved and unresolved comments.

These changes improve the interactivity of the audio player and facilitate user engagement through comments.

+++ app/Livewire/GlobalFileUploader.php
@@ -152,7 +152,7 @@ class GlobalFileUploader extends Component
             $pitch = Pitch::findOrFail($modelId);
 
             // Special handling for client management projects where producer is uploading
-            if ($pitch->project->isClientManagement() && auth()->check() && auth()->id() === $pitch->user_id) {
+            if ($pitch->project->isClientManagement() && auth()->check() && (int) auth()->id() === (int) $pitch->user_id) {
                 // Producer uploading to their own pitch in a client management project
                 $this->fileManagementService->createPitchFileFromS3(
                     $pitch,
